+++ CHANGELOG.md
@@ -4,7 +4,6 @@ It is based on Tidepool Blip 1.27.
 
 ## Unreleased
 ### Added
-- YLP-1088 Rework UX/UI of the YLP footer
 - YLP-1105 Yourloops new blue theme
 - YLP-1106 Switch account from caregiver to hcp should ask for the profession
 - YLP-1122 Patient should enter the team unique ID to accept a team invitation
@@ -13,6 +12,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1217 Patient should be better informed that he should not create an account through web UI
 ### Improvements
 - YLP-22 Print PDF content cleaned and improved (removing useless data)
+- YLP-1088 Rework UX/UI of the YLP footer
 - YLP-1120 Overview tab cleanup
 ### Fixed
 - YLP-1101 Yourloops logo redirect not working for HCPs